Add supports_full_stream_interface
Based on a discussion between PSP and OJB:

We only want one `friend` declaration in each connection; and that is
for `Stream`.

Stream then has a nested Streamable struct, that performs that checks
because it has a right to access private/protected members. It is
implemented as a series of static_asserts that give a helpful error
message indicating what failed exactly. But, that means that when we
instantiate the default-constructed Stream it is effectively a
declaration of intend (i.e., "we intend this connection instance to
support the full Stream interface and hard-fail if it doesn't"). We
cannot perform a "check" if the Connection is even supposed to fulfill
the full Stream interface beforehand. A separate checker would require
to also be friends with all connections..

Hence, up until now we provide a manual hint
("has_full_stream_interface") that can be used to separate the
connections into those that are supposed to fulfill the full stream
interface (and are then used for in-depth hxcomm-tests) and those that
aren't.

Better:
We should have the indication if a Connection is supposed to be
Streamable in a separate type so that Stream only ever gets instantiated
via its default implementation where a failure to fulfil Streamable
should result in a compile error with helpful error message.

This struct is supports_full_stream_interface.

Change-Id: I402df2115373d7becfd3d5a3c76c60b3f1e94c66