quiggeldy: Add at-compile version information
Quiggeldy is planned to be auto-deployed in a nightly fashion. Because
from time to time there are bugfixes/new features added, we need a way
to track in which state quiggeldy was compiled.

This is meant for expert users to ensure a specific version of quiggeldy
is deployed upon debugging.

At a later time, we will add a versioned interface, preventing users
from connecting when versions mismatch.

* Add `-v`/`--version` switch to quiggeldy that prints out with which
  repository at what commit-state it was compiled.
* Add `get_version_string`-method to RCF-interface to allow for remote
  querying of build information.

Depends-On: 13086
Change-Id: I9734f1d1d0214c4953f97e3835534f0e4d93520c