Allow for custom quiggeldy user via env
This change allows clients to specify a custom username when using
quiggeldy: QUIGGELDY_USER_NO_MUNGE

As the name indicates, this will disable munge-based authentication to
quiggeldy and hence requires that munge support is disabled on the
running quiggeldy.

This is useful if several users are emulated by a single unix user, as
was the case at the NICE2021 demo.

Depends-On: 13915

Change-Id: Iadacd1fbf25a57d3c3745bfb26140cd44291181a