quiggeldy: Allow for replayed munge credentials
Munge raises an error if credentials are decoded several times. However,
we usually encode user data that is set once per call. For submitting
reinit programs we perform several actions over the same client. Hence,
we need to ignore this kind of error.
If all calls set user data anew, we could drop this. Realistically this
should be fine though.

Change-Id: I9b1376107cb522bcd90d192b6e8dd562561f6ad7